+++ apps/vscode/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Jupyter: Verify that the ms-toolsai.jupyter extension is installed 
 - Mark temporary .intellisense. files as plaintext so they are not auto-formatted
 - Use triple-dash for block and line comments
+- Visual Editor: Workaround Pandoc 3.1.4 change in `--id-prefix` behavior
 
 ## 1.94.0 (Release on 17 August 2023)